requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
chromadb>=0.5.0
fastapi>=0.111.0
uvicorn>=0.29.0
python-multipart>=0.0.9
opencv-python>=4.9.0